+++ graph_programs/ball2.c
@@ -13,15 +13,15 @@ demonstrate smooth animation.
 
 #define BALLACCEL 5 /* ball acceleration */
 
-int       x, y;
-int       nx, ny;
-int       lx, ly;
-int       xd, yd;
+long      x, y;
+long      nx, ny;
+long      lx, ly;
+long      xd, yd;
 ami_evtrec er;
-int       tc;
-int       cd; /* current display flip select */
-int       ballsize;
-int       halfball;
+long      tc;
+long      cd; /* current display flip select */
+long      ballsize;
+long      halfball;
 
 int chkbrk(void)
 
@@ -39,7 +39,7 @@ int chkbrk(void)
 
 }
 
-void drawball(ami_color c, int x, int y)
+void drawball(ami_color c, long x, long y)
 
 {
